requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
cached-property==1.5.2
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
cramjam==2.5.0
cycler==0.11.0
Cython==0.29.28
dbus-python==1.2.16
debugpy==1.6.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
distro-info==1.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fastparquet==0.8.1
filelock==3.6.0
fonttools==4.33.2
fsspec==2022.3.0
greenlet==1.1.2
httplib2==0.18.1
idna==2.10

ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==4.2.0
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
packaging==21.3
pandarallel==1.6.1
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.1
prometheus-client==0.9.0
prompt-toolkit==3.0.29
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycurl==7.43.0.6
Pygments==2.11.2
PyGObject==3.38.0
pylibacl==0.6.0
pyparsing==3.0.8
pyrsistent==0.15.5
PySimpleSOAP==1.16.2
pystan==2.19.1.1
python-dateutil==2.8.2
python-debian==0.1.39
python-debianbts==3.1.0
python-igraph==0.9.0
pytz==2022.1
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
rdiff-backup==2.0.5
requests==2.25.1
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
sh==1.14.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
stack-data==0.2.0
terminado==0.13.3
texttable==1.6.3
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
urllib3==1.26.5
virtualenv==20.14.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.6.0
zipp==1.0.0
